requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
cachetools==5.5.1; python_version >= '3.7'
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
google-ai-generativelanguage==0.6.15; python_version >= '3.7'
google-api-core[grpc]==2.24.0; python_version >= '3.7'
google-api-python-client==2.159.0; python_version >= '3.7'
google-auth==2.38.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-generativeai==0.8.4; python_version >= '3.9'
googleapis-common-protos==1.66.0; python_version >= '3.7'
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
pillow==11.1.0; python_version >= '3.9'
proto-plus==1.25.0; python_version >= '3.7'
protobuf==5.29.3; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pyparsing==3.2.1; python_version >= '3.1'
python-dotenv==1.0.1; python_version >= '3.8'
python-telegram-bot==21.10; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
sniffio==1.3.1; python_version >= '3.7'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.9'
